Start implementation of better default files

+++ src/impyrium/default_files/__init__.py
@@ -8,7 +8,9 @@ from . import json_file_defaults
 def writeFiles(parentFolder):
     os.makedirs(parentFolder, exist_ok=True)
     for file in DefaultFile.defaultFiles:
-        p = os.path.dirname(f"{parentFolder}/{file.path}")
-        os.makedirs(p, exist_ok=True)
-        with open(f"{parentFolder}/{file.path}", 'w') as f:
-            f.write(file.contents)
+        if file.shouldWrite():
+            p = os.path.dirname(f"{parentFolder}/{file.path}")
+            if not os.path.exists(p):
+                os.makedirs(p, exist_ok=True)
+                with open(f"{parentFolder}/{file.path}", 'w') as f:
+                    f.write(file.contents)
